Make the contact page route people to wikiba.se
The contact page is currently not linked to from the site
However people could possibly still end up here?
So lets just be consistent and point it to the wikiba.se site
which is also what is done in the footer.

I hear this is being touched in more ways soon, but lets just remove
the bit that says to contact me directly...