Use state to handle success and errors
The existing error handling logic seemed to
apply more to the success or failure of dispatching
the store actions rather than the success or failure
of the API calls within those actions.

This commit adjusts that flow to use the success
state. It also adds an additional error state to the store
so that this can be shown the the user (as the existing
code seems to imply it should work).

A feature this code does not cover that was implied
to be there is setting the inFlight state after error.
This means the user will need to reload the page in order
to retry.

This commit also adjusts the mock service worker to mock
a server error.